move testUtils to lib folder
