+++ packages/api/tests/test-endpoints-granules.js
@@ -11,7 +11,7 @@ const models = require('../models');
 const bootstrap = require('../lambdas/bootstrap');
 const granuleEndpoint = require('../endpoints/granules');
 const indexer = require('../es/indexer');
-const { testEndpoint } = require('./testUtils');
+const { testEndpoint } = require('../lib/testUtils');
 const { Search } = require('../es/search');
 
 // create all the variables needed across this test
